% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabstats-package.R
\docType{package}
\name{tabstats}
\alias{tabstats-package}
\alias{tabstats}
\title{tabstats: A Lightweight Toolkit for Displaying Customizable Tables}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A lightweight toolkit that provides functions for printing tables from input data in the R console or terminal with customizable formatting. Supported outputs include American Psychological Association (APA)-style tables (American Psychological Association, 2020, ISBN:9781433832178), correlation matrices, contingency tables, and two-column summary tables.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/s7-stats/tabstats}
  \item Report bugs at \url{https://github.com/s7-stats/tabstats/issues}
}

}
\author{
\strong{Maintainer}: Joshua Marie \email{joshua.marie.k@gmail.com}

}
\keyword{internal}
