% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tabstats-package.R
\docType{package}
\name{tabstats}
\alias{tabstats-package}
\alias{tabstats}
\title{tabstats: Helpers for displaying tables}
\description{
Collection of functions to print the data frames in command line. This package provides helpers to display a statistical results in a data frame, contingency table, correlation matrix, and a summary table in 2 columns form.
}
\author{
\strong{Maintainer}: Joshua Marie \email{joshua.marie.k@gmail.com}

}
\keyword{package}
